+++ src/app/view/mission/new/CopyMissionComponent.js
@@ -1,11 +1,13 @@
 import React, { Component } from 'react';
 import { withStyles } from 'material-ui/styles';
+import { withRouter } from 'react-router-dom';
 import { ContentDuplicate } from 'mdi-material-ui';
 import API from '../../../ctrl/API';
 import Button from 'material-ui/Button';
 import Geosearch from '../../GeosearchComponent';
 import Grid from 'material-ui/Grid';
 import IconGridSelect from '../../IconGridSelectComponent';
+import NewMissionComponent from './NewMissionComponent';
 import SelectList from '../../SelectListComponent';
 import Typography from 'material-ui/Typography';
 import Wait from '../../WaitComponent';
@@ -31,6 +33,82 @@ class CopyMissionComponent extends Component {
 		};
 	}
 	
+	_create() {
+		if(this.state.mission && this.state.mission.id) {
+			if(this.state.place && this.state.place.bbox && this.state.place.names) {
+				PubSub.publish("UI.MESSAGE.WAIT", { message: I18n.t("Copying template information") });
+				
+				//Fetch template details
+				API.GetMissionDetails(this.state.mission.id)
+				.then(m => {
+					const data = NewMissionComponent.MissionToState(m);
+					data.mission.area.name = data.details.areaname = this.state.place.names.join(', ');
+					data.mission.area.bbox = data.datasource.bbox = this.state.place.bbox;
+					if(data.datasource.options && data.datasource.options.geojson) {
+						delete data.datasource.options.geojson;
+					}
+					
+					PubSub.publish("UI.MESSAGE.WAIT", { message: I18n.t("Please wait while the mission is created, it can take a few minutes.") });
+					
+					//Create new mission
+					API.CreateMission(
+						data.mission,
+						data.datasource.source,
+						data.datasource.options,
+						NewMissionComponent.UIEditorsToDb(data),
+						this.props.user.name,
+						this.props.user.id
+					)
+					.then(d => {
+						this._updateCreation(d.id, d.pictoken);
+					})
+					.catch(e => {
+						console.log("Failed creating");
+						console.error(e);
+						PubSub.publish("UI.MESSAGE.WAITDONE");
+						PubSub.publish("UI.MESSAGE.BASIC", { type: "error", message: I18n.t("Oops ! Something went wrong when creating the mission"), details: e.message });
+					});
+				})
+				.catch(e => {
+					console.error(e);
+					PubSub.publish("UI.MESSAGE.WAITDONE");
+					PubSub.publish("UI.MESSAGE.BASIC", { type: "error", message: I18n.t("Oops ! Can't copy mission template"), details: e.message });
+				});
+			}
+			else {
+				PubSub.publish("UI.MESSAGE.BASIC", { type: "alert", message: I18n.t("You have to choose an area for your mission before continuing. Please search for your place's name, then select one of the entry in the result list.") });
+			}
+		}
+		else {
+			PubSub.publish("UI.MESSAGE.BASIC", { type: "alert", message: I18n.t("You have to choose a mission template before continuing. Please click on one of the template in the list.") });
+		}
+	}
+	
+	/**
+	 * Update loading status after creating a mission.
+	 * @private
+	 */
+	_updateCreation(mid, pictoken) {
+		API.GetMissionLoading(pictoken)
+		.then(loading => {
+			if(loading >= 100) {
+				PubSub.publish("UI.MESSAGE.WAITDONE");
+				PubSub.publish("UI.MESSAGE.BASIC", { type: "info", message: I18n.t("Your mission was succesfully created !") });
+				this.props.history.push("/mission/"+mid);
+			}
+			else {
+				PubSub.publish("UI.MESSAGE.WAIT", { progress: loading });
+				setTimeout(() => this._updateCreation(mid, pictoken), 2000);
+			}
+		})
+		.catch(e => {
+			console.log("Failed getting progress");
+			console.error(e);
+			setTimeout(() => this._updateCreation(mid, pictoken), 2000);
+			PubSub.publish("UI.MESSAGE.BASIC", { type: "error", message: I18n.t("Oops ! Something went wrong when creating the mission"), details: e.message });
+		});
+	}
+	
 	render() {
 		let missions = null;
 		
@@ -46,7 +124,7 @@ class CopyMissionComponent extends Component {
 						cols={12}
 						items={THEMES}
 						value={this.state.theme}
-						onChange={id => this.setState({ theme: id })}
+						onChange={id => this.setState({ theme: id, mission: null })}
 					/>
 					
 					{missions && missions.length > 0 ?
@@ -68,13 +146,13 @@ class CopyMissionComponent extends Component {
 			
 			<Grid container spacing={8} alignItems="center" justify="center" style={{marginTop: 10}}>
 				<Grid item xs={12} sm={6} md={4} lg={3}>
-					<Button variant="raised" color="primary" className={this.props.classes.button}>
+					<Button
+						variant="raised"
+						color="primary"
+						className={this.props.classes.button}
+						onClick={() => this._create()}
+					>
 						{I18n.t("Create")}
-					</Button>
-				</Grid>
-				<Grid item xs={12} sm={6} md={4} lg={3}>
-					<Button variant="raised" className={this.props.classes.button}>
-						{I18n.t("Cancel")}
 					</Button>
 				</Grid>
 			</Grid>
@@ -100,9 +178,9 @@ class CopyMissionComponent extends Component {
 		})
 		.catch(e => {
 			console.error(e);
-			PubSub.publish("UI.MESSAGE.BASIC", { type: "error", message: I18n.t("Oops ! Something went wrong when fetching templates")+" "+e.message });
+			PubSub.publish("UI.MESSAGE.BASIC", { type: "error", message: I18n.t("Oops ! Something went wrong when fetching templates"), details: e.message });
 		});
 	}
 }
 
-export default withStyles(styles)(CopyMissionComponent);
+export default withRouter(withStyles(styles)(CopyMissionComponent));